refactor: Info 컴포넌트명 변경, option 추가
