+++ pages/store/[storeId]/info/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Info from "@/components/Info/Info";
+import TableBox from "@/components/TableBox/TableBox";
 import { store } from "@/mocks/mockData/store";
 
 const SELLER_INFO = "판매자 정보";
@@ -8,8 +8,8 @@ const FOOD_INFO = "식품 정보";
 function InfoPage() {
   return (
     <>
-      <Info title={SELLER_INFO} data={store.info.sellerInfo} />
-      <Info title={FOOD_INFO} data={store.info.foodInfo} />
+      <TableBox option="info" title={SELLER_INFO} data={store.info.sellerInfo} />
+      <TableBox option="info" title={FOOD_INFO} data={store.info.foodInfo} />
     </>
   );
 }
